fix: idAfter, totalElements 데이터 타입 Long으로 수정

+++ src/main/java/com/team6/hrbank/dto/changeLog/ChangeLogSearchCondition.java
@@ -10,7 +10,7 @@ public record ChangeLogSearchCondition(
     String ipAddress,
     Instant atFrom,
     Instant atTo,
-    Integer idAfter,
+    Long idAfter,
     String cursor,
     int size,
     String sortField,
